efi_runtime: Copied the structure from userland locally in kernel space
When we have structures that contain pointers, we cannot just do double
dereference on it as in struct->pointer->value. This will end up as a
Page Fault. To fix this problem, I changed efi_runtime kernel module to
first fetch the entire structure into a local copy on stack and use that
to reference the other data members of the struct.

Signed-off-by: Pradeep Gaddam <Pradeep.r.Gaddam@intel.com>